Added documentation to UserBioTableViewCell. Removed the accessToken property since it was not currently in use. Removed the casting of 300 to CGFloat.

+++ MeetupClone/Controllers/ProfileViewController.swift
@@ -34,6 +34,6 @@ extension ProfileViewController: UITableViewDelegate {
         return headerView
     }
     func tableView(_ tableView: UITableView, heightForHeaderInSection section: Int) -> CGFloat {
-        return CGFloat(300)
+        return 300
     }
 }
